(partial) refactor
