+++ .gitignore
@@ -91,3 +91,5 @@ lit-config.json
 
 # Configuration to exclude the extra and local_docs directories
 extra
+packages/plugin-lit/dist/index.js.map
+packages/plugin-lit/dist/index.js
